Add spring-shared module
