+++ inertia4j.spring-boot-4/src/main/java/io/github/inertia4j/springboot4/InertiaSpringAutoconfiguration.java
@@ -1,34 +1,26 @@
 package io.github.inertia4j.springboot4;
 
 import io.github.inertia4j.core.DefaultPageObjectSerializer;
-import io.github.inertia4j.core.SimpleTemplateRenderer;
-import io.github.inertia4j.core.TemplateRenderingException;
 import io.github.inertia4j.spi.PageObjectSerializer;
 import io.github.inertia4j.spi.TemplateRenderer;
-import org.springframework.beans.factory.annotation.Autowired;
+import io.github.inertia4j.springshared.AbstractInertia;
+import io.github.inertia4j.springshared.AbstractInertiaSpringAutoconfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 /**
- * Spring Boot auto-configuration for Inertia4j.
- * Sets up default beans for {@link Inertia}, {@link VersionProvider},
- * {@link PageObjectSerializer}, and {@link TemplateRenderer} if they are not
- * already present in the application context.
+ * Spring Boot 4 implementation of {@link AbstractInertiaSpringAutoconfiguration}.
  */
 @Configuration
-public class InertiaSpringAutoconfiguration {
-    @Autowired
-    InertiaConfigurationProperties properties;
+public class InertiaSpringAutoconfiguration extends AbstractInertiaSpringAutoconfiguration {
+    @Override
+    @Bean
+    @ConditionalOnMissingBean
+    public VersionProvider versionProvider() {
+        return super.versionProvider()::get;
+    }
 
-    /**
-     * Creates the main {@link Inertia} bean if one doesn't already exist.
-     * 
-     * @param versionProvider      The configured or default VersionProvider.
-     * @param pageObjectSerializer The configured or default PageObjectSerializer.
-     * @param templateRenderer     The configured or default TemplateRenderer.
-     * @return The Inertia bean instance.
-     */
     @Bean
     @ConditionalOnMissingBean
     public Inertia inertia(
@@ -39,23 +31,7 @@ public class InertiaSpringAutoconfiguration {
         return new Inertia(versionProvider, pageObjectSerializer, templateRenderer);
     }
 
-    /**
-     * Creates a default {@link VersionProvider} bean that returns "1" if one doesn't already exist.
-     * 
-     * @return A default VersionProvider bean.
-     */
-    @Bean
-    @ConditionalOnMissingBean
-    public VersionProvider versionProvider() {
-        return () -> "1";
-    }
-
-    /**
-     * Creates a default {@link PageObjectSerializer} bean if one doesn't already exist.
-     * Uses Jackson 3 if available on the classpath, otherwise falls back to {@link DefaultPageObjectSerializer}
-     *
-     * @return A default PageObjectSerializer bean.
-     */
+    @Override
     @Bean
     @ConditionalOnMissingBean
     public PageObjectSerializer pageObjectSerializer() {
@@ -65,18 +41,5 @@ public class InertiaSpringAutoconfiguration {
         } catch (ClassNotFoundException e) {
             return new DefaultPageObjectSerializer();
         }
-    }
-
-    /**
-     * Creates a default {@link TemplateRenderer} bean using {@link SimpleTemplateRenderer}
-     * and the template path from {@link InertiaConfigurationProperties} if one doesn't already exist.
-     * 
-     * @return A default TemplateRenderer bean.
-     * @throws TemplateRenderingException if the template file cannot be loaded.
-     */
-    @Bean
-    @ConditionalOnMissingBean
-    public TemplateRenderer templateRenderer() throws TemplateRenderingException {
-        return new SimpleTemplateRenderer(properties.templatePath);
     }
 }
