+++ inertia4j.spring-boot-4/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,2 @@
 io.github.inertia4j.springboot4.InertiaSpringAutoconfiguration
-io.github.inertia4j.springboot4.InertiaConfigurationProperties
+io.github.inertia4j.springshared.InertiaConfigurationProperties
