+++ inertia4j.spring/src/main/java/io/github/inertia4j/spring/VersionProvider.java
@@ -1,12 +1,6 @@
 package io.github.inertia4j.spring;
 
 /**
- * Functional interface for providing the current Inertia asset version.
- * Implementations of this interface can be provided as Spring beans to customize version handling.
+ * Spring Boot 3 implementation of {@link io.github.inertia4j.springshared.VersionProvider}.
  */
-public interface VersionProvider {
-    /**
-     * @return The current Inertia asset version as a String, or {@code null} if versioning is not used.
-     */
-    String get();
-}
+public interface VersionProvider extends io.github.inertia4j.springshared.VersionProvider {}
