+++ inertia4j.spring/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,2 @@
 io.github.inertia4j.spring.InertiaSpringAutoconfiguration
-io.github.inertia4j.spring.InertiaConfigurationProperties
+io.github.inertia4j.springshared.InertiaConfigurationProperties
